Waypoint based AI Car Traffic v2
Waypoint routes completed. Branches script added and branch routes completed. Car prefabs to be used were created. Necessary animation and animation controllers were created. Edits in Waypoint scripts.
