+++ Assets/Convai/Settings/Script State/d1aeb28c-3b92-11ee-8189-42010a40000b/Level001_d1aeb28c-3b92-11ee-8189-42010a40000b_Convai.Scripts.Utils.ConvaiBlinkingHandler_State.data
@@ -1,1 +1,1 @@
-{"faceSkinnedMeshRenderer":{"instanceID":-72764},"indexOfLeftEyelid":-1,"indexOfRightEyelid":-1,"minBlinkDuration":0.1889999955892563,"maxBlinkDuration":0.28999999165534975,"minBlinkInterval":2.0,"maxBlinkInterval":3.0}
+{"faceSkinnedMeshRenderer":{"instanceID":-559362},"indexOfLeftEyelid":-1,"indexOfRightEyelid":-1,"minBlinkDuration":0.1889999955892563,"maxBlinkDuration":0.28999999165534975,"minBlinkInterval":2.0,"maxBlinkInterval":3.0}
