+++ Assets/Convai/Settings/Script State/d1aeb28c-3b92-11ee-8189-42010a40000b/Level001_d1aeb28c-3b92-11ee-8189-42010a40000b_Convai.Scripts.Utils.ConvaiLipSync_State.data
@@ -1,1 +1,1 @@
-{"faceModelType":3,"BlendshapeType":2,"HeadSkinnedMeshRenderer":{"instanceID":-72764},"TeethSkinnedMeshRenderer":{"instanceID":-73200},"TongueSkinnedMeshRenderer":{"instanceID":-73206},"jawBone":{"instanceID":-73038},"tongueBone":{"instanceID":-73046},"tongueBoneOffset":{"x":-0.009999999776482582,"y":0.014999999664723874,"z":0.0},"firstIndex":0,"isCharacterTalking":false}
+{"faceModelType":3,"BlendshapeType":2,"HeadSkinnedMeshRenderer":{"instanceID":-559362},"TeethSkinnedMeshRenderer":{"instanceID":-559798},"TongueSkinnedMeshRenderer":{"instanceID":-559804},"jawBone":{"instanceID":-559636},"tongueBone":{"instanceID":-559644},"tongueBoneOffset":{"x":-0.009999999776482582,"y":0.014999999664723874,"z":0.0},"firstIndex":0,"isCharacterTalking":false}
