+++ Assets/Convai/Settings/Script State/d1aeb28c-3b92-11ee-8189-42010a40000b/Waypoint System_d1aeb28c-3b92-11ee-8189-42010a40000b_Convai.Scripts.Utils.ConvaiHeadTracking_State.data
@@ -1,1 +1,1 @@
-{"targetObject":{"instanceID":45424},"trackingDistance":17.399999618530275,"turnSpeed":6.260000228881836,"bodyLookAtWeight":0.6000000238418579,"headLookAtWeight":0.800000011920929,"eyesLookAtWeight":0.8500000238418579,"lookAway":false}
+{"targetObject":{"instanceID":798620},"trackingDistance":17.399999618530275,"turnSpeed":6.260000228881836,"bodyLookAtWeight":0.6000000238418579,"headLookAtWeight":0.800000011920929,"eyesLookAtWeight":0.8500000238418579,"lookAway":false}
