+++ Assets/Convai/Settings/Script State/d1aeb28c-3b92-11ee-8189-42010a40000b/Waypoint System_d1aeb28c-3b92-11ee-8189-42010a40000b_Convai.Scripts.Utils.ConvaiLipSync_State.data
@@ -1,1 +1,1 @@
-{"faceModelType":3,"BlendshapeType":2,"HeadSkinnedMeshRenderer":{"instanceID":-36170},"TeethSkinnedMeshRenderer":{"instanceID":-36606},"TongueSkinnedMeshRenderer":{"instanceID":-36612},"jawBone":{"instanceID":-36444},"tongueBone":{"instanceID":-36452},"tongueBoneOffset":{"x":-0.009999999776482582,"y":0.014999999664723874,"z":0.0},"firstIndex":0,"isCharacterTalking":false}
+{"faceModelType":3,"BlendshapeType":2,"HeadSkinnedMeshRenderer":{"instanceID":-518976},"TeethSkinnedMeshRenderer":{"instanceID":-519412},"TongueSkinnedMeshRenderer":{"instanceID":-519418},"jawBone":{"instanceID":-519250},"tongueBone":{"instanceID":-519258},"tongueBoneOffset":{"x":-0.009999999776482582,"y":0.014999999664723874,"z":0.0},"firstIndex":0,"isCharacterTalking":false}
