+++ Assets/Downloaded Assets/LowPolyOfficeInterior/LowPolyOffice_2020.3.0_HDRP_v1.0.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 74b1da64efed84c75b531d2319b8aa70
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
