+++ Assets/Downloaded Assets/LowPolyOfficeInterior/LowPolyOffice_2020.3.0_URP_v1.0.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d9828d4144fad4bc1861978c2bcbf360
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
