+++ Assets/Downloaded Assets/PolygonCity/Prefabs/Vehicles/SM_Veh_Car_Muscle_01.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 28f60c12e257c3248a1c5258cc4b1aa9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
