+++ Assets/Downloaded Assets/PolygonNatureBiomes/URP_ExtractMe_Nature_Biomes.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 92f9cd028443b7645a5012a799b57dab
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
