+++ Assets/Prefabs/Vehicles.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 686af44e58adfac4ab9027065ae4bdee
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
