+++ Assets/Prefabs/Vehicles/Vehicle_v1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e45822d7d6ac37e46aebf584b17cc43f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
