+++ Assets/Prefabs/Vehicles/Vehicle_v2.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 09efc52844ed2284cb69d99ec1c365a7
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
