+++ Assets/Prefabs/Vehicles/Vehicle_v3.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c097552ac32e3fc42914e1612bb63313
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
