+++ Assets/Prefabs/Vehicles/Vehicle_v4.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7e1cce02a2fe2bc40be7f49b0b2097ae
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
