+++ Assets/Prefabs/Vehicles/Vehicle_v5.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9fe3e6f1de8923949b9ee9dc100108a4
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
