+++ Assets/Prefabs/Vehicles/Vehicle_v6.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7b4dcf0d8db5cf44b8a126b711ee9fe3
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
