+++ Assets/Prefabs/Vehicles/Vehicle_v7.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ef11b7e44ae166043b480a7d44f7743d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
