+++ Assets/Prefabs/Vehicles/Vehicle_v8.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0bc60f09e6415964b80a109432fa17ce
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
