+++ Assets/Prefabs/Vehicles/animCar.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8faf22e9f1e681d4cb38eef500e2e6a1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
