+++ Assets/Prefabs/Vehicles/tires.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1933ab6d67df75b4a8101dbc79b1179b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
