--- Assets/Scenes/Ozan/CarNavigationController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c6da6463b3dd8c24e95f3d0b6b47e07f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
