--- Assets/Scenes/Ozan/SM_Veh_Car_Muscle_01.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e45822d7d6ac37e46aebf584b17cc43f
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
