--- Assets/Scenes/Ozan/WaypointEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 410e2c882d5798d40a96d1dc6466967f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
