+++ Assets/Scripts/WaypointSystem.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0f48d23f910b9eb4ca1265ebe9d464ad
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
