+++ Assets/Scripts/WaypointSystem/Waypoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 81dda130cdde6e94ebfa8b4b59adc410
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
