+++ Assets/Scripts/WaypointSystem/WaypointManagerWindow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6627dce4cbb0e7740af693622029da00
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
