+++ Assets/Scripts/WaypointSystem/WaypointNavigator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b1903d9ae66ec594e9c6e6cdf48f2e4f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
